+++ src/errors/InteractionsError.js
@@ -4,9 +4,10 @@ const ErrorCodes = require('./ErrorCodes');
 const Messages = require('./Messages');
 
 class InteractionsError extends Error {
-  constructor(code, ...args) {
+  constructor(code, data, ...args) {
     super(message(code, args));
     this.code = code;
+    this.data = data;
     Error.captureStackTrace?.(this, InteractionsError);
   }
 
@@ -23,7 +24,7 @@ class InteractionsError extends Error {
  */
 function message(code, args) {
   if (!(code in ErrorCodes)) {
-    throw new Error('Error code must be a valid DiscordjsErrorCodes');
+    throw new Error('Error code must be a valid DiscordInteractionsErrorCodes');
   }
   const msg = Messages[code];
   if (!msg) throw new Error(`No message associated with error code: ${code}.`);